Create Test Evaluation
Make a program to evaluate multiple choice tests. Each test has 10 questions, each question is worth a point. The first set of data to be read is the correct answers of the test. Then the students’ numbers and their respective answers will be given. The program ends when the number of a given student equals 9999.